--- PokemonProject/App.tsx
@@ -1,11 +1,0 @@
-import React from "react";
-import { Routes } from "./src/routes";
-
-export default function App  () {
-
-  return (
-    
-    <Routes/>
-  
-  );
-}
