+++ PokemonProject/package.json
@@ -9,26 +9,13 @@
     "web": "expo start --web"
   },
   "dependencies": {
-    "@react-native-async-storage/async-storage": "~1.17.3",
-    "@react-navigation/bottom-tabs": "^6.4.0",
-    "@react-navigation/native": "^6.0.14",
-    "@react-navigation/native-stack": "^6.9.1",
-    "@react-navigation/stack": "^6.3.4",
-    "@types/react": "~18.0.24",
-    "@types/react-native": "~0.70.6",
-    "axios": "^1.1.3",
     "expo": "~47.0.6",
-    "expo-linear-gradient": "~12.0.1",
     "expo-status-bar": "~1.4.2",
     "react": "18.1.0",
     "react-native": "0.70.5",
-    "react-native-animatable": "^1.3.3",
-    "react-native-linear-gradient": "^2.6.2",
-    "react-native-paper": "^5.0.0-rc.10",
-    "react-native-safe-area-context": "4.4.1",
-    "react-native-screens": "~3.18.0",
-    "react-native-vector-icons": "^9.2.0",
-    "typescript": "^4.6.3"
+    "typescript": "^4.6.3",
+    "@types/react": "~18.0.24",
+    "@types/react-native": "~0.70.6"
   },
   "devDependencies": {
     "@babel/core": "^7.12.9"
