--- PokemonProject/src/components/BotaoProps/index.tsx
@@ -1,25 +1,0 @@
-import React from "react"
-import { TouchableOpacity, Text, TouchableOpacityProps } from "react-native"
-import { styles } from './styles'
-
-interface BotaoProps extends TouchableOpacityProps {
-    title: string,
-}
-
-export const Botao = ({ title, ...resto }: BotaoProps) => {
-
-
-    return (
-        <TouchableOpacity
-            style={styles.button}
-            {...resto}
-        >
-            <Text
-                style={styles.buttonText}
-            >
-                {title}
-            </Text>
-        </TouchableOpacity>
-    )
-}
-
