--- PokemonProject/src/components/BotaoProps/styles.ts
@@ -1,27 +1,0 @@
-import { StyleSheet } from "react-native";
-
-export const styles = StyleSheet.create({
-    button: {
-        backgroundColor:'#fff',
-        padding: 15,
-        borderRadius: 7,
-        alignItems: 'center',
-        marginTop: 15,
-        letterspacing: 1.5,
-
-    },
-    buttonText: {
-
-        color: '#000',
-        fontSize: 15,
-        fontWeight: 'bold',
-        letterSpacing: 1.5,
-    },
-
-    header:{
-        paddingTop:50,
-        flex:1,
-        
-    }
-
-})
