--- PokemonProject/src/components/BotaoVertical/styles.ts
@@ -1,16 +1,0 @@
-import { StyleSheet } from "react-native";
-
-const styles = StyleSheet.create({
-    text: {
-    
-        color: '#fff'
-    },
-    button: {
-        alignItems: 'center',
-        justifyContent: 'center',
-
-    }
-
-})
-
-export default styles;
