--- PokemonProject/src/components/PokemonCard/styles.ts
@@ -1,24 +1,0 @@
-import { StyleSheet } from "react-native";
-
-export const styles = StyleSheet.create({
-    buttonMagicItem: {
-        backgroundColor: '#1f1e25',
-        padding: 15,
-        borderRadius: 50,
-        alignItems: 'center',
-        marginVertical: 5,
-        flexDirection: "row",
-        justifyContent: "space-between"
-    },
-    textMagicItem: {
-        marginLeft: 5,
-        color: '#FFF',
-        fontSize: 16,
-        width:"80%",
-    },
-    closeIcon: {
-        tintColor: '#fff',
-        height: 25,
-        width: 25
-    },
-})
