--- PokemonProject/src/routes/styles.ts
@@ -1,8 +1,0 @@
-import { StyleSheet } from "react-native";
-
-const styles = StyleSheet.create({
-   
-    
-})
-
-export default styles;
