--- PokemonProject/src/screens/Cart/styles.ts
@@ -1,21 +1,0 @@
-import { StyleSheet, Platform } from "react-native";
-
-export const styles = StyleSheet.create({
-    container: {
-        flex: 1,
-        backgroundColor: '#121015',
-        paddingHorizontal: 30,
-        paddingTop: 60,
-        alignContent: 'space-between'
-    },
-    title: {
-        color: '#fff',
-        fontSize: 24,
-        fontWeight: 'bold',
-    },
-    containerPreco:{
-        marginBottom: 10, 
-        justifyContent:"space-between", 
-        flexDirection: 'row'
-    }
-})
